Feat: add support for SQL Server 2017 in doctrine spatial extension
- Updated composer.json to include mssql2017 configuration for PHPUnit.
- Modified Dockerfile to install Microsoft ODBC driver and dependencies for SQL Server.
- Added new phpunit.mssql2017.xml configuration file for testing.
- Implemented SqlServer platform class and integrated it into AbstractSpatialType.
- Added GeographyType and GeometryType support for SQL Server.
- Created SpSrid DQL function for SQL Server and corresponding tests.
- Enhanced existing tests to support SQL Server platform.

+++ tests/LongitudeOne/Spatial/Tests/Doctrine/ConnectionParameters.php
@@ -78,7 +78,16 @@ class ConnectionParameters
         $connection = DriverManager::getConnection($parameters);
         $manager = $connection->createSchemaManager();
         $dbName = (string) $GLOBALS['db_name'];
-        $manager->dropDatabase($dbName);
+
+        // Drop the database if it exists and create a new one
+        foreach ($manager->listDatabases() as $database) {
+            if ($database === $dbName) {
+                $manager->dropDatabase($dbName);
+
+                break;
+            }
+        }
+
         $manager->createDatabase($dbName);
         $parameters['dbname'] = $dbName;
